+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## unreleased (dd/mm/yyyy)
+
+### Feature
+- O2V-248: adding custom VS ssh port number option
+
+### Fixes
+- O2V-239: fixed handling for destroying an already removed vm
+
 ## v.1.1.3 (02/04/2024)
 
 ### Fixes